Abstract authentication out of Sirius
