+++ src/sirius/mod.rs
@@ -1,57 +1,25 @@
 use std::collections::HashMap;
-
 use anyhow::Result;
 use chrono::{DateTime, Utc};
 use reqwest::{Client, StatusCode};
-use tokio::time::{Duration, Instant};
 
 mod model;
-use model::{AuthResponse, EventResult};
+use model::EventResult;
+mod auth;
+use auth::Auth;
 
 const URL: &str = "https://sirius.fit.cvut.cz/api/v1";
-const OAUTH_URL: &str = "https://auth.fit.cvut.cz/oauth/oauth/token";
 
 #[derive(Debug, Clone)]
 pub struct Sirius {
-    client_id: String,
-    client_secret: String,
-    access_token: Option<String>,
-    expires_in: Instant,
+    auth: Auth,
 }
 
 impl Sirius {
     pub fn new(client_id: String, client_secret: String) -> Sirius {
         Sirius {
-            client_id,
-            client_secret,
-            access_token: None,
-            expires_in: Instant::now(),
+            auth: Auth::new(client_id, client_secret),
         }
-    }
-
-    pub async fn load_access_token(&mut self) -> Result<String> {
-        let now = Instant::now();
-        if self.access_token.is_none() || self.expires_in <= now {
-            // We have to make request
-            let res = Client::new()
-                .post(OAUTH_URL)
-                .form(&[
-                    ("grant_type", "client_credentials"),
-                    ("client_id", &self.client_id),
-                    ("client_secret", &self.client_secret),
-                    ("scope", "cvut:sirius:personal:read"),
-                ])
-                .send()
-                .await?;
-
-            if res.status() == StatusCode::OK {
-                let text = res.text().await?;
-                let content: AuthResponse = serde_json::from_str(&text)?;
-                self.access_token = Some(content.access_token.clone());
-                self.expires_in = Instant::now() + Duration::from_secs(content.expires_in);
-            }
-        }
-        Ok(self.access_token.clone().unwrap())
     }
 
     pub async fn course_events(
@@ -59,7 +27,7 @@ impl Sirius {
         course_code: String,
         options: Options,
     ) -> Result<EventResult> {
-        let token = self.load_access_token().await?;
+        let token = self.auth.get_token().await?;
 
         let mut map: HashMap<String, String> = options.into();
         map.insert("access_token".into(), token);
@@ -67,11 +35,7 @@ impl Sirius {
         let url = format!("{}/courses/{}/events", URL, course_code);
 
         // We have to make request
-        let res = Client::new()
-            .get(url)
-            .form(&map)
-            .send()
-            .await?;
+        let res = Client::new().get(url).form(&map).send().await?;
 
         // TODO: Check for StatusCode::OK
         let text = res.text().await?;
